Now user after register has a token.
